Feature: Extreme changes + working handshakes
Next up is receiving game packets, dispatching them to listeners, and handling reliability internally

+++ src/main/kotlin/raknet/packet/Acknowledge.kt
@@ -33,7 +33,7 @@ data class Record(val isSingle: Boolean, val sequenceNumber: UIntLE, val endSequ
     override fun toString() = "Record(isSingle=$isSingle, sequenceNumber=$sequenceNumber, endSequenceNumber=$endSequenceNumber)"
 }
 
-sealed class Base(id: Short, val recordCount: Short, val record: Record): ConnectedPacket(id) {
+sealed class Base(id: Int, val recordCount: Short, val record: Record): ConnectedPacket(id) {
     override fun encodeOrder(): Array<Any> = arrayOf(recordCount, record)
 }
