+++ src/main/kotlin/raknet/packet/ConnectedPacket.kt
@@ -3,4 +3,4 @@ package raknet.packet
 /**
  * Used as a way to separate unconnected packets from connected packets.
  */
-abstract class ConnectedPacket(id: Short) : DataPacket(id)
+abstract class ConnectedPacket(id: Int) : DataPacket(id)
